Make fancy menu decoration and create file select skeleton

+++ src/draw.c
@@ -95,7 +95,7 @@ void draw_set_font(struct DrawContext* ctx, enum Font font) {
 	}
 }
 
-void draw_text(struct DrawContext* ctx, const char* str, uint8_t x, uint8_t y, struct Pallete pl) {
+void draw_text(struct DrawContext* ctx, const char* str, int16_t x, int16_t y, struct Pallete pl) {
 	struct IRect src = ctx->font_src;
 	uint8_t rootx = src.x;
