+++ src/draw.h
@@ -42,6 +42,6 @@ void draw_rect(struct DrawContext* ctx, struct IRect rect, struct Pallete pl);
 void draw_fill_rect(struct DrawContext* ctx, struct IRect rect, struct Pallete pl);
 
 void draw_set_font(struct DrawContext* ctx, enum Font font);
-void draw_text(struct DrawContext* ctx, const char* str, uint8_t x, uint8_t y, struct Pallete pl);
+void draw_text(struct DrawContext* ctx, const char* str, int16_t x, int16_t y, struct Pallete pl);
 
 #endif // draw_h_INCLUDED
